refactor: extract tryOpenPluginSettings to utils
- Move `tryOpenPluginSettings` to `src/utils/openPluginSettings.ts`
- Update imports in `main.ts` and `src/engine/runTemplateFromFolder.ts`
- Replace `console.error` with the project's internal `log.logError`
- Format code to meet project standards (tab indentation, remove trailing whitespace)
