+++ src/engine/runTemplateFromFolder.ts
@@ -8,6 +8,7 @@ import type ITemplateChoice from "../types/choices/ITemplateChoice";
 import { TemplateChoice } from "../types/choices/TemplateChoice";
 import { normalizeTemplateFolderPaths } from "../utilityObsidian";
 import { isCancellationError, reportError } from "../utils/errorUtils";
+import { tryOpenPluginSettings } from "../utils/openPluginSettings";
 
 export interface RunTemplateFromFolderParams {
 	/** When set, skips the picker and runs this template directly (CLI / scripted). */
@@ -49,28 +50,6 @@ export function hasConfiguredTemplateFolders(plugin: QuickAdd): boolean {
 	return (
 		normalizeTemplateFolderPaths(plugin.settings.templateFolderPaths).length > 0
 	);
-}
-
-/** Opens a plugin's settings tab. Returns false if the internal API is unavailable or throws. */
-export function tryOpenPluginSettings(app: App, pluginId: string): boolean {
-	try {
-		const setting = (
-			app as unknown as {
-				setting?: { open?: () => void; openTabById?: (id: string) => void };
-			}
-		).setting;
-		
-		if (!setting?.open || !setting?.openTabById) {
-			console.error("QuickAdd: Obsidian internal settings API is unavailable.");
-			return false;
-		}
-		setting.open();
-		setting.openTabById(pluginId);
-		return true;
-	} catch (error) {
-		console.error("QuickAdd: Failed to open plugin settings automatically", error);
-		return false;
-	}
 }
 
 function renderTemplateRow(path: string, el: HTMLElement): void {
